% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquiredSpecies.R
\name{acquiredSpecies}
\alias{acquiredSpecies}
\title{Acquisition of Microbial Species}
\usage{
acquiredSpecies(pop,
                parentPop = NULL,
                founderM = NULL,
                sym = NULL,
                rndSeed = NULL,
                globalSP = NULL)
}
\arguments{
\item{pop}{An object of class \code{\link{Pop-class}} generated with \pkg{\link{AlphaSimR}} representing the current population.}

\item{parentPop}{An object of class \code{\link{Pop-class}} generated with \pkg{\link{AlphaSimR}} representing the parental population. If \code{NULL}, the function assumes it needs to compute the microbiota for the base population.}

\item{founderM}{A list containing the details of the microbiota in the founder population, generated by \code{\link{setFounderM}}.}

\item{sym}{A value of 1 to consider the symbiosis effect on microbiota. This value is required when the population is not the base population.}

\item{rndSeed}{A seed value to initialize random sampling.}

\item{globalSP}{An object of class \code{\link{GlobalSP}} containing the global parameters of the simulation. If assigned as \code{gSP}, the function will access it directly.}
}
\value{
A matrix representing the initial microbial species in the current population.
}
\description{
This function determines the initial count of microbial species colonizing an individual, accounting for both the environmental and maternal communities of microorganisms. When saved as \var{acquiredSp}, it can be accessed by other functions.
}
\examples{

# Set global simulation parameters
gSP <- GlobalSP$new(nPop = 1000, nyear = 10, nQTLchr = 100)

# Define the microbiota
gSP$setSpecies(nSpecies = 1000, nSp0 = 600, nSpEff = 100, symbiosis = c(0,1))

# Define the trait
gSP$setTrait(meanP = 10, varP = 3, h2 = 0.10, m2 = 0.10)

# Create founder genetics
founderG <- setFounderG(globalSP = gSP)

# Set simulation genetic parameters
SP <- essentialSP(founder = founderG,
                  minSnpFreq = 0.05, nSnpChr = 10000)

# Create founder microbiota
founderM <- setFounderM(globalSP = gSP)

# Create the founder population
founderPop <- newPop(founderG)

# Create base population
Pop <- randCross(founderPop,
                 nCrosses = gSP$nCross,
                 nProgeny = gSP$nSon)

# Generate acquisition matrix for base population
acquiredSp <- acquiredSpecies(pop = Pop)

}
