% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essentialSP.R
\name{essentialSP}
\alias{essentialSP}
\title{Essential Simulation Parameter Generator}
\usage{
essentialSP(founder = NULL,
            minQTLchr = 1,
            nSnpChr = NULL,
            minSnpChr = NULL,
            minSnpFreq = NULL,
            rndSeed = NULL)
}
\arguments{
\item{founder}{An object of class \code{\link{MapPop-class}} generated by the function \code{\link{setFounderG}}. If \code{NULL}, the function retrieves the founder haplotypes from the system.}

\item{minQTLchr}{A value or vector specifying the minimum number of QTL per chromosome. \emph{Default} is 1.}

\item{nSnpChr}{A value or vector specifying the number of SNPs per chromosome. If \code{NULL}, the \emph{Default} is 100,000 for \dQuote{RABBIT} (based on the 200K Affymetrix Axiom OrcunSNP array \cite{ThermoFisher Scientific}) and 1724 for \dQuote{CATTLE} (based on the Bovine SNP50 v2 BeadChip from \cite{Illumina}). For \dQuote{GENERIC}, a value must be specified.}

\item{minSnpChr}{A value or vector specifying the number of segregating sites for SNPs per chromosome. If \code{NULL}, it is assumed to be half of \var{nSnpChr}.}

\item{minSnpFreq}{The minimum allowable frequency for SNP loci. No minimum SNP frequency is applied if \code{NULL}.}

\item{rndSeed}{A seed to initialize random sampling. If \code{NULL}, a seed is randomly set.}
}
\value{
An object of class \code{\link{SimParam}}.
}
\description{
This function generates an object of type \code{\link{SimParam}} with the essential parameters needed to perform the simulation.
After running this function, other methods from \code{\link{SimParam}} can be used to define additional global parameters not included here.
Saving this object as \var{SP} will allow it to be accessed by other functions with default parameters.
}
\examples{
\dontrun{
# Set genetic parameters
SP <- essentialSP(founder = founderG, nSnpChr = 10000, minSnpFreq = 0.05)
}
}
