% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractM.R
\name{extractM}
\alias{extractM}
\title{Extracting Microbiota Composition}
\usage{
extractM(pop,
                sym,
                globalSP = NULL,
                name = NULL,
                path = "./")
}
\arguments{
\item{pop}{An object of class \code{\link{MapPop-class}}.}

\item{sym}{A value of 1 to consider the symbiosis effect on the microbiota.}

\item{globalSP}{An object of class \code{\link{GlobalSP}}. If \code{NULL}, it will be retrieved from the global environment.}

\item{name}{The name of the \emph{txt} file to be generated.}

\item{path}{The path where the \emph{txt} file will be saved. \emph{Default} is the current directory ("./").}
}
\value{
A matrix representing the microbiota composition.
}
\description{
This function extracts the microbiota composition of a population and generates a \emph{txt} file with the data.
}
\examples{
\dontrun{
microbiota <- extractM(Pop, sym = 0, name = "M_gen3_sym")
}



}
