% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillSp.R
\name{fillSp}
\alias{fillSp}
\title{Fill Species Information}
\usage{
fillSp(pop, w = NULL, mbiome = NULL, sym = 0)
}
\arguments{
\item{pop}{An object of class \code{\link{Pop-class}} generated using the \pkg{\link{AlphaSimR}} package, representing the current population.}

\item{w}{A vector with the \eqn{\omega} effect scaled according to the base population.}

\item{mbiome}{A microbiota abundance matrix generated using the function \code{\link{makeM}}. If \code{NULL}, \code{mbiome} will be extracted from the global environment.}

\item{sym}{A value of 1 to consider the symbiosis effect on the microbiota. \emph{Default} is 0.}
}
\value{
An object of class \code{\link{Pop-class}} with updated microbiota information.
}
\description{
This function incorporates microbial species abundances and the microbiota value into an object of class \code{\link{Pop-class}}.
}
