% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalSP.R
\name{GlobalSP}
\alias{GlobalSP}
\title{Global Simulation Parameters}
\description{
The \sQuote{GlobalSP} class represents global simulation parameters for a specific
simulation scenario. It includes general parameters, trait parameters,
and microbiota parameters.
}
\section{Methods}{


\bold{\emph{Public methods}}

\itemize{
 \item \code{GlobalSP$new()} Create a new GlobalSP object.
 \item \code{GlobalSP$setTrait()} Set trait parameters for the GlobalSP object.
 \item \code{GlobalSP$setSpecies()} Set species parameters for the GlobalSP object.
 \item \code{GlobalSP$getValues()} Retrieve values from the GlobalSP object.
}
}

\examples{
# Set simulation parameters
gSP <- GlobalSP$new(nPop = 1000, nQTLchr = 100, nyear = 10)

# Defining the trait
gSP$setTrait(meanP = 8, varP = 2, h2 = 0.15, m2 = 0.15)

# Defining the microbiota
gSP$setSpecies(nSpecies = 1000, nSp0 = 600, nSpEff = 35)

}
\references{
Coyte KZ, Rakoff-Nahoum S. Understanding competition and cooperation within the mammalian gut microbiome. Curr Biol. 2019;29:R538-R544
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nPop}}{(numeric)
Number of individuals in the founder population.}

\item{\code{dataM}}{(TXT file)
Real microbiota data (count data). TXT file separated by tabulations. Indicate the complete path if the file is not in the working directory}

\item{\code{nChr}}{(numeric)
Total number of chromosomes to simulate using \code{runcMacs} from \code{\link{AlphaSimR}}.}

\item{\code{nQTLchr}}{(numeric)
Number of QTL per chromosome affecting the simulated trait.}

\item{\code{segSITESchr}}{(numeric)
Total number of segregating sites to simulate using \code{runcMacs} from \code{\link{AlphaSimR}}.
If it is \code{NULL}, all are retained.}

\item{\code{nt}}{(numeric)
If OpenMP is available, this allows for simulating chromosomes in parallel. \emph{Default} is 1.}

\item{\code{nSim}}{(numeric)
Number of repetitions of the simulation. \emph{Default} is 1.}

\item{\code{animal}}{(character)
Type of animal. Currently, three species histories are available: \dQuote{GENERIC}, \dQuote{CATTLE},
and \dQuote{RABBIT}. All except the rabbit demographic histories are implemented in \pkg{\link{AlphaSimR}}.
The rabbit demographic is implemented in this package following \cite{Carneiro et al. (2014)}.}

\item{\code{selType}}{(character)
Selection type (e.g., \dQuote{Divergent}, \dQuote{Low}, \dQuote{High}).}

\item{\code{model}}{(character)
Simulation model. There are six models implemented:
\itemize{
 \item \dQuote{G}: Effect of the genome only
 \item \dQuote{M}: Effect of microbiota only
 \item \dQuote{NMH}: Hologenome without microbial heritability (Null Microbial heritability)
 \item \dQuote{LMH}: Hologenome considering Low Microbial Heritability
 \item \dQuote{MHM}: Hologenome considering Medium Microbial Heritability
 \item \dQuote{HMH}: Hologenome considering High Microbial Heritability
 \item \dQuote{H}: Hologenome considering variability of Microbial heritability
}
All models can run simultaneously starting from the same base population if \dQuote{All} is specified.}

\item{\code{nyear}}{(numeric)
Number of years.}

\item{\code{nSire}}{(numeric)
Number of sires.}

\item{\code{nDam}}{(numeric)
Number of dams.}

\item{\code{nCross}}{(numeric)
Number of crosses.}

\item{\code{nSon}}{(numeric)
Number of progeny per cross.}

\item{\code{scale}}{(character)
Type of transformation to perform in the data. \emph{Default} is log ()
Centered-log ratio transformation is also allowed ("clr")}

\item{\code{meanP}}{(numeric)
Mean value for the simulated trait.}

\item{\code{varP}}{(numeric)
Variance of the simulated trait.}

\item{\code{h2}}{(numeric)
Heritability of the simulated trait.}

\item{\code{m2}}{(numeric)
Microbiability for the simulated trait.}

\item{\code{nSpecies}}{(numeric)
Total number of microbial species simulated in a community.}

\item{\code{nSp0}}{(numeric)
Total number of simulated microbial species in the base population.}

\item{\code{nSpEff}}{(numeric)
Number of microbial species affecting the simulated trait.}

\item{\code{PM}}{(numeric)
Proportion of the initial parental microbiota (dam microbiota) exposed to the offspring.}

\item{\code{EM}}{(numeric)
Proportion of the initial environmental microbiota exposed to the offspring.}

\item{\code{propMH}}{(numeric)
Proportion of microbial species affected by the host genome.}

\item{\code{propQTL}}{(numeric)
Proportion of QTLs affecting the abundance of the microbial species.}

\item{\code{propMH.wSp}}{(numeric)
Proportion of microbial species with effects on the simulated trait that are affected by the host genome.}

\item{\code{MH.G}}{(numeric)
Microbial heritability in the G scenario.}

\item{\code{MH.M}}{(numeric)
Microbial heritability in the M scenario.}

\item{\code{MH.low}}{(numeric)
Microbial heritability in the LMH scenario.}

\item{\code{MH.medium}}{(numeric)
Microbial heritability in the MHM scenario.}

\item{\code{MH.high}}{(numeric)
Microbial heritability in the HMH scenario.}

\item{\code{MH.H}}{(numeric vector)
Vector of variable microbial heritability}

\item{\code{meanMH}}{(numeric)
Mean of microbial heritability distribution}

\item{\code{varMH}}{(numeric)
Variance of microbial heritability}

\item{\code{symbiosis}}{(numeric)
Value or vector to specify whether to simulate (1) or not (0) the microbial species interaction or symbiosis effect on microbial abundance.
There are five types of interactions simulated according to \cite{Coyte KZ and Rakoff-Nahoum S (2019)}: Commensalism, Mutualism, Ammensalism, Competition, and Exploitation/Pathogen.}

\item{\code{s2}}{(numeric)
Proportion of microbial species abundance due to the symbiosis or interaction between microbial species.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GlobalSP-new}{\code{GlobalSP$new()}}
\item \href{#method-GlobalSP-setTrait}{\code{GlobalSP$setTrait()}}
\item \href{#method-GlobalSP-setSpecies}{\code{GlobalSP$setSpecies()}}
\item \href{#method-GlobalSP-getValues}{\code{GlobalSP$getValues()}}
\item \href{#method-GlobalSP-getPrivate}{\code{GlobalSP$getPrivate()}}
\item \href{#method-GlobalSP-clone}{\code{GlobalSP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSP-new"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSP-new}{}}}
\subsection{Method \code{new()}}{
Function to initialize the parameters needed for the simulation.
Saving this object as \var{gSP} will facilitate the functionality of the other functions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSP$new(
  nPop,
  dataM = NULL,
  nChr = 1,
  nQTLchr,
  segSITESchr = NULL,
  animal = "GENERIC",
  nSire = 25,
  nDam = 125,
  nCross = NULL,
  selType = "Divergent",
  nyear,
  model = "H",
  nt = 1,
  nSim = 1,
  scale = "log"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nPop}}{The total number of individuals to simulate in the founder population.}

\item{\code{dataM}}{Path to TXT file format of microbiota composition (count data).}

\item{\code{nChr}}{The total number of chromosomes to simulate using \code{runcMacs} from \code{\link{AlphaSimR}}. \emph{Default} is 1.}

\item{\code{nQTLchr}}{A value or vector indicating the number of QTL per chromosome affecting the simulated trait.}

\item{\code{segSITESchr}}{The total number of segregating sites per chromosome to simulate using \code{runcMacs} from \code{\link{AlphaSimR}}.
If it is \var{NULL}, all are retained.}

\item{\code{animal}}{The species to simulate. Can be \dQuote{GENERIC}, \dQuote{CATTLE}, or \dQuote{RABBIT}.
All except the rabbit demographic histories are implemented in \pkg{\link{AlphaSimR}}.}

\item{\code{nSire}}{Number of breeding sires. \emph{Default} is 25.}

\item{\code{nDam}}{Number of breeding dam. \emph{Default} is 125.}

\item{\code{nCross}}{Number of crosses per sire. \emph{Default} in \var{NULL}. If it is \var{NULL}, it will directly computed.}

\item{\code{selType}}{The selection type (e.g., \dQuote{Divergent}, \dQuote{Low}, \dQuote{High}).}

\item{\code{nyear}}{The total number of years to simulate.}

\item{\code{model}}{The simulation model. \emph{Default} H model.
There are five models implemented: \dQuote{G}, \dQuote{M}, \dQuote{NMH}, \dQuote{LMH}, \dQuote{MHM}, and \dQuote{HMH}.}

\item{\code{nt}}{If OpenMP is available, this allows for simulating chromosomes in parallel. \emph{Default} is 1.}

\item{\code{nSim}}{The number of repetitions of the simulation. \emph{Default} is 1.}

\item{\code{scale}}{Transformation to perform in the data. \emph{Default} log() is taken.}

\item{\code{nSon}}{Number of progeny per cross. \emph{Default} is 8.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSP-setTrait"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSP-setTrait}{}}}
\subsection{Method \code{setTrait()}}{
Function to set trait parameters for the GlobalSP object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSP$setTrait(meanP, varP, h2, m2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meanP}}{The mean value for the trait.}

\item{\code{varP}}{The variance of the trait.}

\item{\code{h2}}{Heritability of the trait.}

\item{\code{m2}}{Microbiability of the trait.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSP-setSpecies"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSP-setSpecies}{}}}
\subsection{Method \code{setSpecies()}}{
Function to set species parameters for the GlobalSP object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSP$setSpecies(
  nSpecies = NULL,
  nSp0 = NULL,
  nSpEff,
  PM = 0.5,
  EM = 0.5,
  propMH = 0.1,
  propQTL = 0.1,
  propMH.wSp = 0.5,
  symbiosis = c(0, 1),
  s2 = 0.2,
  MH.G = 0.2,
  MH.M = 0.2,
  MH.low = 0.1,
  MH.medium = 0.3,
  MH.high = 0.6,
  MH.H = NULL,
  meanMH = NULL,
  varMH = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nSpecies}}{Total number of microbial species in the community. \emph{Default} is NULL}

\item{\code{nSp0}}{Number of initial microbial species. \emph{Default} is NULL}

\item{\code{nSpEff}}{Number of effective microbial species affecting the trait.}

\item{\code{PM}}{Proportion of parental microbiota exposed to offspring.}

\item{\code{EM}}{Proportion of environmental microbiota exposed to offspring.}

\item{\code{propMH}}{Proportion of species influenced by the host genome.}

\item{\code{propQTL}}{Proportion of QTLs affecting the microbial species.}

\item{\code{propMH.wSp}}{Proportion of species affecting the trait influenced by the genome.}

\item{\code{symbiosis}}{Whether to simulate symbiotic effects (1) or not (0).}

\item{\code{s2}}{Proportion of microbial abundance explained by symbiosis.}

\item{\code{MH.G}}{Microbial heritability in the G scenario.}

\item{\code{MH.M}}{Microbial heritability in the M scenario.}

\item{\code{MH.low}}{Microbial heritability in the LMH scenario.}

\item{\code{MH.medium}}{Microbial heritability in the MHM scenario.}

\item{\code{MH.high}}{Microbial heritability in the HMH scenario.}

\item{\code{MH.H}}{Vector of microbial heritability in H scenario.}

\item{\code{meanMH}}{Mean of microbial heritability in H scenario.}

\item{\code{varM}}{Variance of microbial heritability in H scenario.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSP-getValues"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSP-getValues}{}}}
\subsection{Method \code{getValues()}}{
Function to retrieve the values of the GlobalSP object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSP$getValues()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSP-getPrivate"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSP-getPrivate}{}}}
\subsection{Method \code{getPrivate()}}{
Function to retrieve private parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSP$getPrivate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalSP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalSP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalSP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
