% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeP.R
\name{makeP}
\alias{makeP}
\title{Compute Phenotypes}
\usage{
makeP(pop,
      model,
      sym = 0,
      limTrait = NULL,
      sex = "both",
      globalSP = NULL,
      rndSeed = NULL)
}
\arguments{
\item{pop}{An object of \code{\link{MapPop-class}} for computing the phenotype.}

\item{model}{A string indicating the model to simulate. Only the values \dQuote{G}, \dQuote{M}, \dQuote{NMH}, \dQuote{LMH}, \dQuote{MMH}, or \dQuote{HMH} are available.}

\item{sym}{A value of 1 to consider the symbiosis effect on microbiota. \emph{Default} is 0.}

\item{limTrait}{A vector with the minimum and maximum values for the trait. If \code{NULL}, it will be obtained from the global environment.}

\item{sex}{Required if the phenotype is attributed to only one sex (e.g., litter size). Use \dQuote{M} for male and \dQuote{F} for female. Default is \dQuote{both} for both sexes.}

\item{globalSP}{An object of class \code{\link{GlobalSP}}. If \code{NULL}, it will be obtained from the global environment.}

\item{rndSeed}{A seed for random sampling. If it is \code{NULL}, a random seed will be set.}
}
\value{
An object of \code{\link{MapPop-class}} that includes the computed phenotypes.
}
\description{
This function computes the phenotype of a population according to a specified scenario and generation.
}
\examples{
# Set simulation global parameters
gSP <- GlobalSP$new(nPop = 1000, nyear = 10, nQTLchr = 100)

# Define the microbiota
gSP$setSpecies(nSpecies = 1000, nSp0 = 600, nSpEff = 100, symbiosis = c(0, 1))

# Define the trait
gSP$setTrait(meanP = 10, varP = 3, h2 = 0.10, m2 = 0.10)

# Create founder genetic
founderG <- setFounderG(globalSP = gSP)

# Set simulation genetic parameters
SP <- essentialSP(founder = founderG, minSnpFreq = 0.05, nSnpChr = 10000)

# Create founder microbiota
founderM <- setFounderM(globalSP = gSP)

# Create the founder population
founderPop <- newPop(founderG)

# Create the base population
basePop <- simBasePop(model = "NMH")

# Compute phenotype of the base population
Pop <- makeP(pop = basePop$Pop, model = "NMH", sym = 1, limTrait = c(0, 18), sex = "F")
}
