% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSelection.R
\name{makeSelection}
\alias{makeSelection}
\title{Perform Selection}
\usage{
makeSelection(g = NULL,
              model = NULL,
              selType = NULL,
              sym = NULL,
              limTrait = NULL,
              sex = "both",
              LS = FALSE,
              globalSP = NULL,
              save = c(),
              iter = 1,
              progressBar = FALSE)
}
\arguments{
\item{g}{Number of generations for selection. If \code{NULL}, it will be obtained from the global environment.}

\item{model}{A value for the model to simulate. Only values according to \dQuote{G}, \dQuote{M}, \dQuote{NMH}, \dQuote{LMH}, \dQuote{MMH}, \dQuote{HMH}, \dQuote{All} are available. If \code{NULL}, it will be obtained from the global environment.}

\item{selType}{Selection type (e.g., \dQuote{Divergent}, \dQuote{Low}, \dQuote{High}). If \code{NULL}, it will be obtained from the global environment.}

\item{sym}{A value of 1 to consider the symbiosis effect on microbiota. If \code{NULL}, it will be obtained from the global environment.}

\item{limTrait}{Vector with the minimum and maximum values for the trait. If \code{NULL}, it will be obtained from the global environment.}

\item{sex}{Indicates if the phenotype belongs exclusively to females ("F"). Default is \dQuote{both}.}

\item{LS}{\code{TRUE} if the phenotype being evaluated is litter size (LS). Default is \code{FALSE}.}

\item{globalSP}{An object of class \code{\link{GlobalSP}}. If \code{NULL}, it will be obtained from the global environment.}

\item{save}{If not null, the base and last population of the simulation will be saved.}

\item{iter}{Number of iterations to perform. If \code{NULL}, just one iteration will be done.}

\item{progressBar}{If TRUE, a progress bar and text will be displayed. \emph{Default} is \code{FALSE}.}
}
\value{
A list of class \code{retValues} with the following information:\enumerate{
  \item{\strong{Generation}}: Generation computed
  \item{\strong{Model}}: Model computed
  \item{\strong{iter}}: Iteration performed. In case the selection will be replicated
  \item{\strong{P}}: Average of the population phenotype
  \item{\strong{varP}}: Variance of the phenotype
  \item{\strong{gv}}: Average of the additive genetic value
  \item{\strong{varG}}: Genetic variance
  \item{\strong{S_gv}}: Selection differential for additive genetic values
  \item{\strong{mv}}: Average of the microbiome values
  \item{\strong{varM}}: Variance of the microbiome values
  \item{\strong{S_mv}}: Selection differential for the microbiome values
  }
}
\description{
This function automates the selection process across different generations and returns a report of class \code{retValues}.
}
\examples{
# Set simulation global parameters
gSP = GlobalSP$new(nPop = 1000, nyear = 10, nQTLchr = 100)

# Define the microbiota
gSP$setSpecies(nSpecies = 1000, nSp0 = 600, nSpEff = 100, symbiosis = c(0,1))

# Define trait
gSP$setTrait(meanP = 10, varP = 3, h2 = 0.10, m2 = 0.10)

# Create founder genetic
founderG = setFounderG(globalSP = gSP)

# Set simulation genetic parameters
SP <- essentialSP(founder = founderG,
                  minSnpFreq = 0.05, nSnpChr = 10000)

# Create founder microbiota
founderM <- setFounderM(globalSP = gSP)

# Create the founder population
founderPop <- newPop(founderG)

# Initialize the selection process
selection <- makeSelection(sex = "F", LS = TRUE, limTrait = c(0,18))
}
