% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matingPLAN.R
\name{matingPLAN}
\alias{matingPLAN}
\title{Create Mating Plan}
\usage{
matingPLAN(parent,
           nCross = NULL,
           selType = NULL,
           globalSP = NULL)
}
\arguments{
\item{parent}{List of selected breeding animals obtained using the function \code{\link{selectBreeding}}.}

\item{nCross}{Number of females mating with each male. If \code{NULL}, it will be obtained from the global environment.}

\item{selType}{Selection type (e.g., \dQuote{Divergent}, \dQuote{Low}, \dQuote{High}). If \code{NULL}, it will be obtained from the global environment.}

\item{globalSP}{An object of class \code{\link{GlobalSP}}. If \code{NULL}, it will be obtained from the global environment.}
}
\value{
A list including two dataframes with the mating plans for increasing and decreasing the phenotype:\enumerate{
  \item{\strong{PopLow}}: Mating plan for decreasing the phenotype. ID of breeding females (Dam) and males (Sire), and value for the genetic or phenotypic value (Sel_value).
  \item{\strong{PopHigh}}: Mating plan for increasing the phenotype. ID of breeding females (Dam) and males (Sire), and value for the genetic or phenotypic value (Sel_value).
  }
}
\description{
Creates crosses based on the selected breeding animals using \code{\link{selectBreeding}}.

Selection of breeding animals is according to the Phenotype (\dQuote{phe}) or the direct genetic value (\dQuote{gv}).
}
\examples{

# Set simulation global parameters
gSP = GlobalSP$new(nPop = 1000, nyear = 10, nQTLchr = 100)

# Define the microbiota
gSP$setSpecies(nSpecies = 1000, nSp0 = 600, nSpEff = 100, symbiosis = c(0,1))

# Define trait
gSP$setTrait(meanP = 10, varP = 3, h2 = 0.10, m2 = 0.10)

# Create founder genetic
founderG = setFounderG(globalSP = gSP)

# Set simulation genetic parameters
SP <- essentialSP(founder = founderG,
                  minSnpFreq = 0.05, nSnpChr = 10000)

# Create founder microbiota
founderM <- setFounderM(globalSP = gSP)

# Create the founder population
founderPop <- newPop(founderG)

# Create the base population
basePop <- simBasePop(model = "NMH")

# Compute phenotype in a base population
Pop <- makeP(pop = basePop$Pop, model = "NMH", sym = 1, limTrait = c(0,18), sex = "F")

# Select breeding animals from basePop
parent <- selectBreeding(pop = Pop, nDam = 125, nSire = 25, sex = "F", g0 = TRUE)

# Create mating plan for next generation
crossPlan <- matingPLAN(parent = parent)

}
