% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotREND.R
\name{plotREND}
\alias{plotREND}
\title{Plot Trend}
\usage{
plotREND(data,
                plot = "line",
                draw = "Model",
                model = NULL,
                effect = "P",
                type_se = "classic",
                data2 = NULL,
                selType = "High",
                name = NULL,
                path = "./",
                show = FALSE,
                dpi = 600,
                width = 17,
                height = 12,
                units = "cm")
}
\arguments{
\item{data}{A data.frame from the \code{retValues} object.}

\item{plot}{The type of plot to generate (\dQuote{boxplot} or \dQuote{line}). The default is \dQuote{line}, which creates trends for all effects within a single model when \code{draw} is set to \dQuote{Model} (the \code{model} parameter is required). To plot trends of all models for a specific effect, use the \dQuote{Effect} option in \code{draw}. The default is \dQuote{P}, which plots the phenotype for all evaluated models. To change it, use the \code{effect} argument with \dQuote{gv} or \dQuote{mv}. For boxplot, just data is required.}

\item{draw}{To graph all the effect for a specific model (\dQuote{Model}; requires \code{model} argument) or all models for a specific effect (\dQuote{Effect}; requires \code{effect} argument). Default \dQuote{Model}}

\item{model}{The model for which to generate the plot line for \code{draw} option equal to \dQuote{Model}.}

\item{effect}{The effect to plot when \code{plot} is set to \dQuote{Effect}. The default is \dQuote{P} for Phenotype. Other options are \dQuote{gv} for genetic values and \dQuote{mv} for microbiota values.}

\item{type_se}{Graph style for plotting the standard error (\dQuote{ribbon} or \dQuote{classic}). Default is \dQuote{classic}.}

\item{data2}{A data.frame from the \code{retValues} object. Required if \code{selType} is set to \dQuote{Divergent}.}

\item{selType}{The population to plot. The default is \dQuote{High}.}

\item{name}{The name of the graph file.}

\item{path}{The path where the graph file will be saved.}

\item{show}{If TRUE, the plot will be displayed in the current R session.}

\item{dpi}{The resolution of the image. The default is 600 dpi.}

\item{width}{The width of the image. The default is 17 cm.}

\item{height}{The height of the image. The default is 12 cm.}

\item{units}{The units for the width and height.}
}
\value{
A graph file and a \code{\link{ggplot2}} object.
}
\description{
\code{plotREND} generates a linear plot of genetic, phenotypic, and microbiome trends across generations for one scenario or for all, focusing on a single effect.
}
\examples{
\dontrun{
plotREND(data=selection[["Report"]][["PopHigh"]]$Symbiosis,
         plot = "boxplot",
         selType = "High",
         show = TRUE,
         name = "lastG_boxplot_symbiosis_high.tiff")
}
}
