% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retValues.R
\name{retValues}
\alias{retValues}
\title{Retrieve Simulation Values}
\usage{
retValues(model,
          g,
          sym = 0,
          selType,
          iter = 1,
          pop = NULL,
          breedInd = NULL)
}
\arguments{
\item{model}{The model value to simulate. Only values corresponding to \dQuote{G}, \dQuote{M}, \dQuote{NMH}, \dQuote{LMH}, \dQuote{MMH}, or \dQuote{HMH} are available.}

\item{g}{Current generation of selection. A value is required.}

\item{sym}{A value of 1 to consider the symbiosis effect on microbiota. \emph{Default} is 0.}

\item{selType}{Selection type (\dQuote{Low}, \dQuote{High}). A value is required.}

\item{iter}{Current iteration. \emph{Default} is 1.}

\item{pop}{An object of class \code{\link{MapPop-class}}. If \code{NULL}, it will be obtained from the environment (\code{Pop}).}

\item{breedInd}{A list of selected breeding animals obtained using the function \code{\link{selectBreeding}}. If \code{NULL}, it will be obtained from the environment (\code{parent}). A list of type \code{selectBreeding} can be used.}
}
\value{
A list containing two sub-lists with data frames that include the above information for the populations with high (\dQuote{PopHigh}) and low (\dQuote{PopLow}) phenotypes. Each population has a data frame for values with (\dQuote{Symbiosis}) and without (\dQuote{No_symbiosis}) the symbiosis effect.
}
\description{
Retrieve information from the simulation process, including:
\enumerate{
  \item{\strong{Generation}}: Computed generation
  \item{\strong{Model}}: Computed model
  \item{\strong{iter}}: Iteration performed; relevant if the selection is replicated
  \item{\strong{P}}: Average phenotype of the population
  \item{\strong{varP}}: Phenotypic variance
  \item{\strong{gv}}: Average additive genetic value
  \item{\strong{varG}}: Genetic variance
  \item{\strong{S_gv}}: Mean genetic value of the breeding animals
  \item{\strong{mv}}: Average microbiota value
  \item{\strong{varM}}: Variance of microbiota value
  \item{\strong{S_mv}}: Mean microbiota value of the breeding animals
  }
}
\examples{

#Set simulation global parameters
gSP = GlobalSP$new(nPop = 1000, nyear = 10, nQTLchr = 100)

#Defining the microbiota
gSP$setSpecies(nSpecies = 1000, nSp0 = 600, nSpEff = 100, symbiosis = c(0,1))

#Defining trait
gSP$setTrait(meanP = 10,varP=3,h2=0.10,m2=0.10)

#Create founder genetic
founderG = setFounderG(globalSP = gSP)

#Set simulation genetic parameters
SP <- essentialSP(founder = founderG,
                  minSnpFreq = 0.05,rndSeed=NULL, nSnpChr = 10000)

#Create founder microbiota
founderM <- setFounderM(globalSP = gSP)

#Create the founder population
founderPop <- newPop(founderG)

#Create the base population
basePop <- simBasePop(model = "NMH")

#Computing phenotype in a base population
Pop <- makeP(pop = basePop$Pop, model = "NMH", sym = 0, limTrait = c(0,18), sex = "F")

#Selection breeding animals from basePop
parent <- selectBreeding(pop = Pop, nDam = 125, nSire = 25, sex = "F", g0 = TRUE)

#Retrieving values
report <- retValues(model = "NMH", g = 1, sym = 0, selType = "Low", pop=Pop, breedInd= parent)
}
