% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePop.R
\name{savePop}
\alias{savePop}
\title{Save Population}
\usage{
savePop(pop, model, g, save = NULL)
}
\arguments{
\item{pop}{An object of class \code{\link{MapPop-class}}.}

\item{model}{The model used for the simulation. Valid values are \dQuote{G}, \dQuote{M}, \dQuote{NMH}, \dQuote{LMH}, \dQuote{MMH}, or \dQuote{HMH}.}

\item{g}{The current generation of selection. A value is required.}

\item{save}{A list of previously saved populations. If a variable called \emph{sortedPop} exists, it will be retrieved from the environment and used to continue storing the populations.}
}
\value{
A list of \code{\link{MapPop-class}} objects, ordered by generation.
}
\description{
This function allows populations to be stored in a common list.
}
\examples{
\dontrun{
# Saving the population
storedPop <- savePop(Pop, model = "NMH")
}

}
