% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectBreeding.R
\name{selectBreeding}
\alias{selectBreeding}
\title{Selection of Breeding Animals}
\usage{
selectBreeding(pop,
               nDam = NULL,
               nSire = NULL,
               selType = NULL,
               selby = "phe",
               sex = "both",
               maxFS = 2,
               g0 = FALSE,
               LS = FALSE,
               globalSP = NULL)
}
\arguments{
\item{pop}{an object of \code{\link{MapPop-class}}}

\item{nDam}{number of females to select. If \code{NULL}, it will be set by \code{nDam} from \code{\link{GlobalSP}}. Default 125.}

\item{nSire}{number of males to select. If \code{NULL}, it will be set by \code{nSire} from \code{\link{GlobalSP}}. Default 25.}

\item{selType}{Selection type (e.g., \dQuote{Divergent}, \dQuote{Low}, \dQuote{High}). If \code{NULL}, it will be obtained from the global environment}

\item{selby}{type of selection; Phenotypic selection (\dQuote{phe}; Default) or Genomic selection (\dQuote{gv})}

\item{sex}{indicate if the phenotype belongs exclusively to females ("F"). Default \dQuote{both}}

\item{maxFS}{maximum number of full siblings allowed when selecting females.}

\item{g0}{\code{TRUE} for indicating that the animals will be selected from the base population. Default \code{FALSE}}

\item{LS}{\code{TRUE} if the phenotype being evaluated is litter size (LS). Default \code{FALSE}}

\item{globalSP}{an object of class \code{\link{GlobalSP}}. If \code{NULL}, it will be obtained from the global environment.}
}
\value{
a list including four dataframes:\enumerate{
  \item{\strong{Dam_low}}: ID of breeding females for decreasing the phenotype, mother ID, father ID and value for the genetic value or phenotype, depending on the selby
  \item{\strong{Dam_high}}: ID of breeding females for increasing the phenotype, mother ID, father ID and value for the genetic value or phenotype, depending on the selby
  \item{\strong{Male_low}}: ID of breeding males for decreasing the phenotype, mother ID, and father ID
  \item{\strong{Male_high}}: ID of breeding males for increasing the phenotype, mother ID, and father ID
  }
}
\description{
Selection of breeding animals according to the Phenotype (\dQuote{phe}) or the direct genetic value (\dQuote{gv}).
}
\examples{
# Set simulation global parameters
gSP = GlobalSP$new(nPop = 1000, nyear = 10, nQTLchr = 100)

# Defining the microbiota
gSP$setSpecies(nSpecies = 1000, nSp0 = 600, nSpEff = 100, symbiosis = c(0,1))

# Defining trait
gSP$setTrait(meanP = 10, varP = 3, h2 = 0.10, m2 = 0.10)

# Create founder genetic
founderG = setFounderG(globalSP = gSP)

# Set simulation genetic parameters
SP <- essentialSP(founder = founderG,
                  minSnpFreq = 0.05, nSnpChr = 10000)

# Create founder microbiota
founderM <- setFounderM(globalSP = gSP)

# Create the founder population
founderPop <- newPop(founderG)

# Create the base population
basePop <- simBasePop(model = "NMH")

# Computing phenotype in a base population
Pop <- makeP(pop = basePop$Pop, model = "NMH", sym = 1, limTrait = c(0,18), sex = "F")

# Selection of breeding animals from basePop
parent <- selectBreeding(pop = Pop, nDam = 125, nSire = 25, sex = "F", g0 = TRUE)

}
