% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFounderG.R
\name{setFounderG}
\alias{setFounderG}
\title{Generate Founder Population Demography}
\usage{
setFounderG(globalSP = NULL, rndSeed = NULL)
}
\arguments{
\item{globalSP}{An object of class \code{\link{GlobalSP}} containing the global parameters of the simulation. If assigned as \code{gSP}, the function will access it directly.}

\item{rndSeed}{A seed for random sampling. If it is \code{NULL}, a random seed will be set.}
}
\value{
An object of class \code{\link{MapPop-class}}.
}
\description{
This function calls \code{\link{runMacs}} or \code{\link{runMacs2}} from the \pkg{\link{AlphaSimR}} package to generate the founder haplotypes. For more details, refer to the documentation for those functions.
This function allows the implementation of the rabbit's demographic history using \code{\link{runMacs2}}, based on \cite{Carneiro et al. (2014)}. Saving this object as \var{founderGenomes} will enable access by default in other functions.
}
\examples{
# Set global simulation parameters
gSP <- GlobalSP$new(nPop = 1000, nyear = 10, nQTLchr = 100)

# Define the microbiota
gSP$setSpecies(nSpecies = 1000, nSp0 = 600, nSpEff = 100, symbiosis = c(0,1))

# Define the trait
gSP$setTrait(meanP = 10, varP = 3.8, h2 = 0.15, m2 = 0.15)

# Create founder genetics
founderG <- setFounderG(globalSP = gSP)

}
\references{
Carneiro M, Albert FW, Afonso S, Pereira RJ, Burbano H, Campos R, et al. The genomic architecture of population divergence between subspecies of the European rabbit. PLoS Genet. 2014;10:e1003519
}
