% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFounderM.R
\name{setFounderM}
\alias{setFounderM}
\title{Generate Microbiota Architecture}
\usage{
setFounderM(globalSP = NULL)
}
\arguments{
\item{globalSP}{an object of \code{\link{GlobalSP}} with the global parameters of the simulation. If it is \code{NULL}, the globalSP will be extracted from the Global Environment}
}
\value{
a list including three dataframes:\enumerate{
  \item{\strong{architecture}: Dataframe with the architecture of the microbiota summarized in the following columns:\itemize{
       \item{\code{RA_EM}: Mean of the relative abundance of the environmental microbiota}
       \item{\code{Species}: Artificial name for microbial species}
       \item{\code{EM}: Mean of the real abundance of the environmental microbiota}
       \item{\code{PM0}: Vector indicating which microbial species will form the microbiota of the base population}
       \item{\code{PM}: Mean of the real abundance of the parental microbiota}
       \item{\code{GR_PM}: Mean of the growth rate of the microbial species when belonging to the parental microbiota}
       \item{\code{SD}: Standard deviation of the growth rate}
       \item{\code{RA_PM0}: Mean of the relative abundance of the microbial species in the parental microbiota}
       \item{\code{w}: Vector of the effect of species abundance on the trait}
       \item{\code{beta}: Vector of 0 or 1 indicating the species abundance influenced by the host genome}}}
   \item{\strong{beta}: Dataframe with the values of the genetic effect on the abundance of each species. Dimensions \eqn{nQTL} x \eqn{nSpecies}}
   \item{\strong{symbiosis}: Dataframe with the symbiosis effect among species. Dimensions \eqn{nSpecies} x \eqn{nSpecies}}}
}
\description{
Generation of the architecture for the parental and environmental microbiota at founder population. The \pkg{\link{mobsim}} package is used to help simulate the species abundance distribution. Saving it as \code{founderM} will facilitate access by other functions.
}
\details{
The simulation of \eqn{\beta} and \eqn{\omega} is based on \cite{Perez-Enciso et al. (2021)}.
\eqn{\beta} is the genetic effect of the QTLs on the species abundance, which follows a gamma distribution \eqn{\Gamma(shape = 0.2, scale=5)} as suggested in \cite{Caballero et al. (2015)}. \eqn{\omega} is the effect of the species abundance on the phenotype which follows a gamma distribution \eqn{\Gamma(shape = 1.4, scale=3.8)} (\cite{Pérez-Enciso et al., 2021}). The hyperparameters of this distribution consider that few microbial species have an effect on the phenotype (\cite{Duvallet et al., 2017}; \cite{Casto-Rebollo et al., 2023}).
}
\examples{

# Set simulation parameters
gSP <- GlobalSP$new(nPop = 1000, nQTLchr = 100, nyear = 10)

# Defining the microbiome
gSP$setSpecies(nSpecies = 1000, nSp0 = 600, nSpEff = 35)

# Making microbiome architecture
founderM <- setFounderM(globalSP = gSP)

}
\references{
Pérez-Enciso M, Zingaretti LM, Ramayo-Caldas Y et al. Opportunities and limits of combining microbiome and genome data for complex trait prediction. Genet Sel Evol. 2021;53:65. \doi{10.1186/s12711-021-00658-7}

Caballero A, Tenesa A, Keightley PD. The nature of genetic variation for complex traits revealed by GWAS and regional heritability mapping analyses. Genetics. 2015;201:1601–13. \doi{10.1534/genetics.115.177220}

Duvallet C, Gibbons SM, Gurry T et al. Meta-analysis of gut microbiome studies identifies disease-specific and shared responses. Nat Commun 2017;8:1784. \doi{10.1038/s41467-017-01973-8}
}
