% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySel.R
\name{summarySel}
\alias{summarySel}
\title{Summary of a generation of selection}
\usage{
summarySel(data, g)
}
\arguments{
\item{data}{data.frame from \code{retValues} object.}

\item{g}{generation for calculating the statistics}
}
\value{
a data.frame with the following columns: \enumerate{
  \item{\strong{Generation}}: Generation computed
  \item{\strong{Model}}: Scenario computed
  \item{\strong{Gv_mean}}: Model computed
  \item{\strong{Gv_sd}}: Iteration performed. In case the selection will be replicated
  \item{\strong{Gv_range}}: Average of the population phenotype
  \item{\strong{Gvba_mean}}: Mean of the microbiota value for the breeding animals
  \item{\strong{Gvba_sd}}: Standard deviation of the microbiota value for the breeding animals
  \item{\strong{Gvba_range}}: Range of the microbiota value for the breeding animals
  \item{\strong{Mv_mean}}: Variance of the phenotype
  \item{\strong{Mv_sd}}: Average of the additive genetic value
  \item{\strong{Mv_range}}: Genetic variance
  \item{\strong{Mvba_mean}}: Mean of the microbiota value for the breeding animals
  \item{\strong{Mvba_sd}}: Standard deviation of the microbiota value for the breeding animals
  \item{\strong{Mvba_range}}: Range of the microbiota value for the breeding animals
  \item{\strong{P_mean}}: phenotypic mean
  \item{\strong{P_sd}}: standard deviation of the phenotype
  \item{\strong{P_range}}: range of the phenotype
  \item{\strong{Pba_mean}}: Mean of the phenotypic value for the breeding animals
  \item{\strong{Pba_sd}}: Standard deviation of the phenotypic value for the breeding animals
  \item{\strong{Pba_range}}: Range of the phenotypic value for the breeding animals
  }
}
\description{
Extract the mean, standard deviation and range of genetic, microbiota and phenotypic valuesof the scenarios, taking into account all iterations performed.
}
\examples{
\dontrun{
summary_lastg <- summarySel(selection$Report$PopHigh$Symbiosis, globalSP$nyear)
}
}
